api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
ultralytics>=8.4.0
opencv-python-headless>=4.6.0
Pillow>=9.0.0
numpy>=1.23.0
